# Trace threadlocal entry is never removed, only nulled