@carterkozak 2022-02-17
Hmm, benchmark data isn't helpful because it's using an MDC map implementation which allocates. Using `-Dlog4j2.garbagefreeThreadContextMap=true` we get:

Before
```
Benchmark                                                              (observability)  Mode  Cnt     Score      Error   Units
TracingBenchmark.traceWithSingleSpan                                            SAMPLE  avgt    3   347.288 ±   47.650   ns/op
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate                             SAMPLE  avgt    3  1317.659 ±  179.190  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate.norm                        SAMPLE  avgt    3   560.023 ±    0.029    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space                    SAMPLE  avgt    3  1296.084 ± 1778.947  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space.norm               SAMPLE  avgt    3   550.672 ±  685.866    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space                SAMPLE  avgt    3     0.005 ±    0.019  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space.norm           SAMPLE  avgt    3     0.002 ±    0.008    B/op
TracingBenchmark.traceWithSingleSpan:·gc.count                                  SAMPLE  avgt    3    23.000             counts
TracingBenchmark.traceWithSingleSpan:·gc.time                                   SAMPLE  avgt    3    26.000                 ms
TracingBenchmark.traceWithSingleSpan                                     DO_NOT_SAMPLE  avgt    3   163.359 ±    6.863   ns/op
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate                      DO_NOT_SAMPLE  avgt    3   920.359 ±   38.607  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate.norm                 DO_NOT_SAMPLE  avgt    3   184.008 ±    0.015    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space             DO_NOT_SAMPLE  avgt    3   901.539 ± 1780.372  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space.norm        DO_NOT_SAMPLE  avgt    3   180.275 ±  363.839    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space         DO_NOT_SAMPLE  avgt    3     0.008 ±    0.169  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space.norm    DO_NOT_SAMPLE  avgt    3     0.002 ±    0.034    B/op
TracingBenchmark.traceWithSingleSpan:·gc.count                           DO_NOT_SAMPLE  avgt    3    16.000             counts
TracingBenchmark.traceWithSingleSpan:·gc.time                            DO_NOT_SAMPLE  avgt    3    19.000                 ms
TracingBenchmark.traceWithSingleSpan                                         UNDECIDED  avgt    3   176.826 ±   18.086   ns/op
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate                          UNDECIDED  avgt    3   868.412 ±   89.363  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate.norm                     UNDECIDED  avgt    3   187.935 ±    0.235    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space                 UNDECIDED  avgt    3   845.183 ±    0.200  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space.norm            UNDECIDED  avgt    3   182.912 ±   18.705    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space             UNDECIDED  avgt    3     0.007 ±    0.151  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space.norm        UNDECIDED  avgt    3     0.002 ±    0.032    B/op
TracingBenchmark.traceWithSingleSpan:·gc.count                               UNDECIDED  avgt    3    15.000             counts
TracingBenchmark.traceWithSingleSpan:·gc.time                                UNDECIDED  avgt    3    17.000                 ms
```

After
```
Benchmark                                                              (observability)  Mode  Cnt     Score      Error   Units
TracingBenchmark.traceWithSingleSpan                                            SAMPLE  avgt    3   313.471 ±   67.044   ns/op
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate                             SAMPLE  avgt    3  1376.501 ±  295.484  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate.norm                        SAMPLE  avgt    3   528.017 ±    0.025    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space                    SAMPLE  avgt    3  1403.523 ± 2335.557  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space.norm               SAMPLE  avgt    3   538.074 ±  783.768    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space                SAMPLE  avgt    3     0.009 ±    0.174  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space.norm           SAMPLE  avgt    3     0.003 ±    0.066    B/op
TracingBenchmark.traceWithSingleSpan:·gc.count                                  SAMPLE  avgt    3    19.000             counts
TracingBenchmark.traceWithSingleSpan:·gc.time                                   SAMPLE  avgt    3    21.000                 ms
TracingBenchmark.traceWithSingleSpan                                     DO_NOT_SAMPLE  avgt    3   111.628 ±   37.147   ns/op
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate                      DO_NOT_SAMPLE  avgt    3  1112.801 ±  370.462  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate.norm                 DO_NOT_SAMPLE  avgt    3   152.007 ±    0.009    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space             DO_NOT_SAMPLE  avgt    3  1126.925 ± 1779.805  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space.norm        DO_NOT_SAMPLE  avgt    3   153.824 ±  200.768    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space         DO_NOT_SAMPLE  avgt    3     0.009 ±    0.160  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space.norm    DO_NOT_SAMPLE  avgt    3     0.001 ±    0.022    B/op
TracingBenchmark.traceWithSingleSpan:·gc.count                           DO_NOT_SAMPLE  avgt    3    20.000             counts
TracingBenchmark.traceWithSingleSpan:·gc.time                            DO_NOT_SAMPLE  avgt    3    22.000                 ms
TracingBenchmark.traceWithSingleSpan                                         UNDECIDED  avgt    3   138.785 ±   46.854   ns/op
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate                          UNDECIDED  avgt    3   918.204 ±  310.462  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.alloc.rate.norm                     UNDECIDED  avgt    3   155.932 ±    0.193    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space                 UNDECIDED  avgt    3   901.572 ± 1781.446  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Eden_Space.norm            UNDECIDED  avgt    3   153.125 ±  301.941    B/op
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space             UNDECIDED  avgt    3     0.009 ±    0.112  MB/sec
TracingBenchmark.traceWithSingleSpan:·gc.churn.G1_Survivor_Space.norm        UNDECIDED  avgt    3     0.001 ±    0.018    B/op
TracingBenchmark.traceWithSingleSpan:·gc.count                               UNDECIDED  avgt    3    16.000             counts
TracingBenchmark.traceWithSingleSpan:·gc.time                                UNDECIDED  avgt    3    19.000                 ms
```