@svc-autorelease 2022-02-17
Released [6.8.0](https://github.com/palantir/tracing-java/releases/tag/6.8.0)